feat(milk): separate types from the kafka client
